requirements.txt
requests>=2.28.0
python-dotenv>=0.21.0
Pillow>=9.0.0
openai>=0.27.0  # Optional for AI features
pandas>=2.0.0   # For Excel reading
openpyxl>=3.0.0 # Excel file support
xlrd>=2.0.0     # For older .xls files